refactored links and nar bar, removed containers and updated myaccount
