+++ .travis.yml
@@ -24,4 +24,4 @@ deploy:
   provider: heroku
   app: mad-hatters # see README
   api_key:
-    secure: "fElz/fmvuSFEjGTnljmCwG4AS205gdf9ThiE4E5NFieM7Mco2jgxVD0r42R5apPKFmW0XWDj1S41T6AlKk5GjVLSI9QFV8PIwJo5K4R78qTpA1SXGbHyMy5yChtphzLGzZsjrdQxtNyVn+XHli9lLUU2Zo3hWpkd80wWc7LCeeuTEcd1iSJ5auPMAzPxPuRvx5lrdzKTnjbWpgvPms2WHyssCrjWZUXhq2PxgVCi9e6l/feLOAj5m5O1Vh3+jgUUJeLqvc4ZfNdtldDUsAFOsuAcEUhTog+MhdUkWGxwFGsgm39GiNiDoAuZOPSri1sGvVMVyGLTzgbvYqij1UXowUau0fb8nmwia10R9q3vi+T13jWNko4cI7aDQrKps3HiMpr4nnL6Ui5rXbSUA19eO44WIowRfxwvNbJWeJKCKahxLQGv/rksTY1lVheW2HIjcoPqdxlBWkYRBdGtC3yl+Gzj1OQamcwEkwhIaqw2dMPkW7GFjSKCOy/UrjfT/k/jhpRq2rXgvAccaHyZKYFflTrVBpCornBX1TVl4v2u6fJ6GREC4FhJ4cxAEfWch421uUddU7iewf39yDk8b42TBW3AmEIbdNYlV4Z9TYFi0Odwc1LSAjXSrgJEZfA+zfiuEtLFZ6JTJl7n26rBYa4QpQQAmhI6Blx+X3uR1GPy8sE="
+    secure: "c2HyKJ9WlZnfJEkx2pl+PkkOOMYBlxE0CYOTdKCyj3+WHZdRCe9qxCRHlG8Iw2kwhteAv3Y+9h8PhPn8LCGua9/IFCvJWYcxibVRHazh9CtbmERF+QmDRusGnLDoExvElVyV+Wnl+sgCKSFr5uPSeqLuEBNlT4FkLlua5Cpk0LhhRIIxURsL1oQEhIqJxXhc/KrBVedTsket6Fb8OdEQoTBS7mVbDIn2BzPLUxEXdy2XAbKcRf1t+VKSI03XvAt+mwr2UFmVY+75vlxhOi2jX0ah0vAur8yW9ypNYNQbWER4Pyg5FiFUyqeUNxRsvn2o3UGt94UjQZrSMYzCY8cBUnK7twmx82Zf4gavSq41Fnh0YOre4yI6BHxBZkw1r7gHbIxs3309yXocUe87OoTRdsJdWQWMWtzs3l8yr7gmURZNf8flKzn0i5sGIZEZIexH8O8AzEwDx4xawOlC1NfWesuUCN73ByjHdrjRkVJWvEdmGQFOwT8JBeu7urs7g65DGXW618v6Zor4Ujls0kNqli94peQrWorUjuDiss+kaJBm+ufF456UeP8/p+SyH92b5CbH8rh+I41bO/Q3XJ+mx0r/Ewttz6CXZS20pwLFfR6NS/e5o36jRWysrOfKd57aRKWc6WAXf+zo33FevM2P2hYQLmMv3ZLzw/YK4EFd5H0="
