+++ client/routes.js
@@ -4,11 +4,11 @@ import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import {Login, Signup, UserHome} from './components'
 import {me} from './store'
-import AllHats from './components/AllHatsContainer'
-import SingleHat from './components/SingleHatContainer'
+import AllHats from './components/AllHats'
+import SingleHat from './components/SingleHat'
 import Welcome from './components/Welcome'
 import userHome from './components/user-home'
-import AdminPage from './components/AdminPage'
+import MyAccount from './components/MyAccount'
 
 /**
  * COMPONENT
@@ -33,7 +33,7 @@ class Routes extends Component {
           <Switch>
             {/* Routes placed here are only available after logging in */}
             <Route path="/home" component={userHome} />
-            <Route path="/admin" component={AdminPage} />
+            <Route path="/myaccount" component={MyAccount} />
           </Switch>
         )}
         {/* Displays our Login component as a fallback */}
